backend/backend/requirements.txt
appdirs==1.4.3
attrs==19.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coverage==5.1
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
flask-redis==0.4.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib
more-itertools==8.2.0
msgpack==0.6.2
numpy
pandas
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
Pillow==8.0.1
progress==1.5
py==1.8.1
pyparsing==2.4.6
pytest==5.4.2
pytest-cov==2.8.1
pytoml==0.1.21
redis==3.5.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
validators==0.15.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
xlrd